contact page done, about page changes, mail stuff
The contact page is complete, as far as I can see. I still need to test
it in the production environment, but the mail templating is certainly
working. I made some changes in the JavaScript as part of this work.

I have also added more text to the about page. I think I will still add
a photo of Rhiain (the one used on Twitter), and update the licesning
information accordingly.

There have been several styling changes too.