shop, renaming
I've worked a fair bit on the shop page. The initial landing part is
done, AFAIK, but I have yet to make a functional (heh) basket/checkout.
In all likelihood, the basket will just be attached to a message on the
contact page or something.

I have also renamed a load of files due to my own previous ineptitude. I
have also added a "not found" thumbnail SVG, just in case.

A lot of time was spend today fiddling with the database, which doesn't
show here.